Extract 'manage folders' screen to ManageFoldersFragment
